Update DetectParalogs to accept a phylogeny
Now accepting a newick phylogeny to make plot with tips ordered according to it. Will also make a tanglegram comparing the phylogeny to the paralogy-based clusters.

New R dependencies: dendextend, gplots, ape.

note that missingness is coded as -100 when clustering paralogy rate so as to use it as info, and to avoid missing data clustering issues. This seems reasonable considering that lack of capture is probably a function of divergence from the ingroup, so it just has the effect of ensuring that the outgroups are clustered as such.